Inline the path slicer module into __init__.py
